third_party/blink/web_tests/virtual/css-custom-state-only-deprecated-syntax/external/wpt/custom-elements/state/state-css-selector-expected.txt

This is a testharness.js-based test.
[FAIL] state selector influences has() when state is applied
  assert_equals: expected "rgb(0, 255, 255)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
